Add potential fix for jar file not running
See: https://github.com/nus-cs2103-AY2223S1/forum/issues/169
