+++ src/main/java/roger/Main.java
@@ -15,32 +15,10 @@ import roger.ui.controllers.MainWindow;
 /**
  * A GUI for Duke using FXML.
  */
-public class Main extends Application {
+public class Main {
 
-    private Parser parser;
-    private Storage storage;
-    private Roger roger;
-
-    /**
-     * Start the Roger backend, and the JavaFX UI elements.
-     * @param stage
-     */
-    @Override
-    public void start(Stage stage) {
-        parser = new Parser();
-        storage = new Storage(Paths.get("data/database.txt"));
-        roger = new Roger(parser, storage);
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            MainWindow mainWindow = fxmlLoader.<MainWindow>getController();
-            mainWindow.setRogerClass(roger);
-            stage.show();
-            mainWindow.greet();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+    public static void main(String[] args) {
+        Launcher.launch(Application.class, args);
+        System.exit(0);
     }
 }
